requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>1.0.0
pypdf
streamlit
faiss-cpu
python-dotenv
